requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django==2.2
django-allauth==0.52.0
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.11.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.17.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==22.0
Pillow==9.4.0
pipdeptree==2.3.3
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
boto3==1.26.59
django-storages<1.13.2